# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AxionPAY Payment API
  version: 0.2.0
servers:
  - url: http://localhost:3060
...
security:
  - ApiKeyAuth: []
    BearerAuth: []
paths:
...
  /webhooks/pix:
    post:
      tags: [Webhooks]
      summary: PIX webhook
      security: []
      parameters:
        - name: x-pix-signature
          in: header
          required: false
          schema:
            type: string
        - name: x-webhook-signature
          in: header
          required: true
          schema:
            type: string
        - name: x-webhook-timestamp
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PixWebhookRequest"
      responses:
        "200":
          description: Payment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /webhooks/pagarme:
    post:
      tags: [Webhooks]
      summary: Provider webhook (Pagar.me/Woovi payload)
      security: []
      parameters:
        - name: x-woovi-signature
          in: header
          required: false
          schema:
            type: string
        - name: x-pagarme-signature
          in: header
          required: false
          schema:
            type: string
        - name: x-hub-signature
          in: header
          required: false
          schema:
            type: string
        - name: x-webhook-signature
          in: header
          required: true
          schema:
            type: string
        - name: x-webhook-timestamp
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Webhook processed
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /webhooks/woovi:
    post:
      tags: [Webhooks]
      summary: Woovi webhook (alias)
      security: []
      parameters:
        - name: x-woovi-signature
          in: header
          required: false
          schema:
            type: string
        - name: x-webhook-signature
          in: header
          required: true
          schema:
            type: string
        - name: x-webhook-timestamp
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Webhook processed
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [ok, message]
      properties:
        ok: { type: boolean, example: false }
        error: { type: string }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
...
    PaymentMethod:
      type: string
      enum: [pix, card]
    PaymentStatus:
      type: string
      enum: [pending, authorized, paid, failed, canceled, refunded, expired]
...
    PixWebhookRequest:
      type: object
      required: [providerReference, event]
      properties:
        providerReference:
          type: string
        event:
          type: string
          enum: [PIX_CONFIRMED, PIX_FAILED, PIX_EXPIRED]
    Transaction:
      type: object
      properties:
        id: { type: string }
        amount: { type: number }
        amount_cents: { type: integer }
        currency: { type: string }
        method:
          $ref: "#/components/schemas/PaymentMethod"
        status:
          $ref: "#/components/schemas/PaymentStatus"
        capture: { type: boolean }
        customer:
          type: object
          nullable: true
        provider:
          type: string
          nullable: true
        providerReference:
          type: string
          nullable: true
        methodDetails:
          type: object
          nullable: true
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        events:
          type: array
          items:
            $ref: "#/components/schemas/PaymentEvent"
    PaymentEvent:
      type: object
      properties:
        type: { type: string }
        at:
          type: string
          format: date-time
    PaymentResponse:
      type: object
      properties:
        ok: { type: boolean }
        transaction:
          $ref: "#/components/schemas/Transaction"
...